+++ port/stm32f0/ow_port_f0.h
@@ -26,7 +26,7 @@
 #ifndef OW_PORT_F0_H
 #define OW_PORT_F0_H
 
-#include "onewire.h"
+#include "onewire_internal.h"
 #include "ow_bits.h"
 #include "stm32f0xx.h"