+++ port/stm32f1/ow_port_f1.h
@@ -23,7 +23,7 @@
 #ifndef OW_PORT_F1_H
 #define OW_PORT_F1_H
 
-#include "onewire.h"
+#include "onewire_internal.h"
 #include "ow_bits.h"
 #include "stm32f1xx.h"